# gateway/fixtures/mobile.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soccer mobile companion API",
    "version": "1.0.0"
  },
  "paths": {
    "/mobile/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_mobile_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Mobile Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/mobile/demo": {
      "get": {
        "summary": "Demo",
        "operationId": "demo_mobile_demo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Demo Mobile Demo Get"
                }
              }
            }
          }
        }
      }
    },
    "/mobile/predict": {
      "post": {
        "summary": "Predict",
        "operationId": "predict_mobile_predict_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobilePredictIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePrediction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MobilePredictIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode",
            "default": "demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "MobilePredictIn"
      },
      "MobilePrediction": {
        "properties": {
          "service": {
            "type": "string",
            "const": "soccer",
            "title": "Service",
            "default": "soccer"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "pending_approval"
            ],
            "title": "Status"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "default": ""
          },
          "prediction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction"
          },
          "degraded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded"
          },
          "tool_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "approval_request": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Request"
          },
          "provenance": {
            "$ref": "#/components/schemas/MobileProvenance"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "status",
          "thread_id",
          "provenance"
        ],
        "title": "MobilePrediction"
      },
      "MobileProvenance": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "synthetic-fixture",
              "gateway"
            ],
            "title": "Kind"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "data_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Backend"
          },
          "fixture_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Id"
          },
          "fixture_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Sha256"
          },
          "generated_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At Utc"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "MobileProvenance"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}